[BE] FIX: length -> size 변경 & totalLength -> totalPage

+++ backend/src/main/java/org/ftclub/cabinet/dto/BlockedUserPaginationDto.java
@@ -12,5 +12,5 @@ import lombok.Getter;
 public class BlockedUserPaginationDto {
 
 	private final List<UserBlockedInfoDto> result;
-	private final Long totalPage;
+	private final Integer totalPage;
 }
